restgears/requirements.txt
Django==1.3
PIL==1.1.7
django-piston==0.2.2
docutils==0.7
wsgiref==0.1.2
